Added sign out feature
-Polished sign in and create account
-sign out feature
-Auto update and clear JWT
